% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alt_params}
\alias{alt_params}
\title{Alternative Specific Parameters}
\usage{
alt_params(N_individuals, mu_alt, var_cov)
}
\arguments{
\item{N_individuals}{Number of individuals}

\item{mu_alt}{A vector of mean(s) of the target parameter(s) to be assigned.}

\item{var_cov}{Variance-Covariance Matrix of the parameters}
}
\description{
This function aims at assigning random or deterministic, correlated or uncorrelated parameters of corresponding alternatives to individuals.
}
\examples{
alt_params(N_individuals = 100, mu_alt = c(2, 3, 4, 5), var_cov = matrix(c(4, 2, 3, 2)))

}
