% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{binary_characteristics}
\alias{binary_characteristics}
\title{Socioeconomic Data, Gender and Habit}
\usage{
binary_characteristics(n, p)
}
\arguments{
\item{n}{Number of observations to be generated.}

\item{p}{The probability of success}
}
\description{
Based on Bernoulli's principle, data related to gender and bio-habit are generated. In order for this function to provide us with correlated features and more precise and realistic data, the function needs to be improved and apply more advanced sampling methods.
}
\examples{
binary_characteristics(n = 100, p = 0.5)

}
