% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{choice}
\alias{choice}
\title{Decision Making Process Based on RUM}
\usage{
choice(
  total_utility,
  Gumbel = c(0, 1),
  no_of_alternatives_per_choiceset = 2,
  no_choice = TRUE
)
}
\arguments{
\item{total_utility}{A matrix of utilities rendered by total_utility function of the package.}

\item{Gumbel}{Specifications of the noise generator. This noise is assigned to the no-choice alternative if no_choice = TRUE}

\item{no_of_alternatives_per_choiceset}{Number of alternatives in each choice set without considering no-choice alternative.}

\item{no_choice}{If TRUE, no-choice alternative will be added to the choice sets, and a random noise will be compared with utilities.}
}
\description{
This function reveals the choices made by artificial individuals based on the fact that the alternative with the highest utility is the favorable choice of a rational individual.
}
\examples{
choice(total_utility(see the arguments), Gumbel = c(0, 1), no_of_alternatives_per_choiceset = 2, no_choice = TRUE)

}
