% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{continuous_characteristics}
\alias{continuous_characteristics}
\title{Socioeconomic Data, Age and Income Features}
\usage{
continuous_characteristics(n, mu, sigma, lowerbound, upperbound, feature)
}
\arguments{
\item{n}{Number of observations to be synthesized}

\item{mu}{The mean of the hypothetical distribution(Here, normal dist.)}

\item{sigma}{Standard deviation or true deviation(if known) of the distribution}

\item{lowerbound}{The lowest amount of the range in which data must be limited}

\item{upperbound}{The highest amount of the range in which data must be limited}

\item{feature}{If it is equal to one, the income data will be labelled.}
}
\description{
The function strives to generate continuous variables based on theoretical distributions. One of the aspects of this function which needs to be enhanced is to enable user to enter his/her desired theoretical distributions and their specifications.
}
\examples{
continuous_characteristics(n = 100, mu = 2.14, lowerbound = 0, upperbound = +inf, feature = 1)

}
