% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dataset_generation}
\alias{dataset_generation}
\title{Dataset Generation}
\usage{
dataset_generation(
  design,
  input2,
  N_individuals,
  mu,
  mu_alt,
  var_cov,
  interactions,
  Gumbel,
  no_of_alternatives_per_choiceset = 2,
  no_choice = TRUE
)
}
\arguments{
\item{design}{The design made by one of the design generator functions in this package.}

\item{input2}{A list of success probabilities for gender and bio-habit features, mu, sigma, lowerbounds and higherbounds for age and income features.}

\item{N_individuals}{Number of synthetic individuals to be generated.}

\item{mu}{A vector of the mean of the constant parameters to be assigned.}

\item{mu_alt}{A vector of mean(s) of the target parameter(s) to be assigned.}

\item{var_cov}{Variance-Covariance Matrix of the parameters.}

\item{interactions}{A vector keeping the order of columns of the design for which you wish to have interaction among them.}

\item{Gumbel}{Specifications of the noise generator. It is a vector containing the mean and beta parameters of gumbel distribution.}

\item{no_of_alternatives_per_choiceset}{Number of alternatives in each choice set without considering no-choice alternative.}

\item{no_choice}{If TRUE, no-choice alternative will be added to the choice sets, and a random noise will be compared with utilities.}
}
\description{
This function facilitates data generation process by gathering all of the functions and inputs together. Finally, dplyr functions will help to extract more features such as unobserved utilities, or it decodes and put labels on features.
}
