% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{FFD}
\alias{FFD}
\title{Full Factorial Design}
\usage{
FFD(input = NULL)
}
\arguments{
\item{input}{This is a list containing attribute levels(attr.lvls), continuous varibles' levels(c.lvls) if any exists, and a vector of the type of variables, in which D stands for dummy and C stands for continuous variables.}
}
\description{
This function provides a full factorial design composed of continuous and binary variables.
}
\examples{
input = list(attr.lvls = c(7, 2, 2),c.lvls = list(c(1.5, 2, 2.5, 3, 3.5, 4, 4.5)),type = c("C", "D", "D"))
 FFD(input)

}
