% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{individual}
\alias{individual}
\title{Main Data Generation Function}
\usage{
individual(input2, N_individuals)
}
\arguments{
\item{input2}{A list of success probabilities for gender and bio-habit features, mu, sigma, lowerbounds and higherbounds for age and income features}

\item{N_individuals}{Number of observations to be produced}
}
\description{
This function facilitates the generation of hypothetical individuals' data
}
\examples{
input2 = list(bpar = c(0.49, 0.35),musigma = list( c(mu = 1220,sigma = 2150,lowerbound = 0,upperbound = 1000000000000,feature = 1),c(mu = 39.74,sigma = 18.89,lowerbound = 18,upperbound = 85,feature = 2)))
individual(input2, N_individuals = 100)
}
