% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{paper_design}
\alias{paper_design}
\title{Paper Design Imitation}
\usage{
paper_design(input3, samples, continuous_levels)
}
\arguments{
\item{input3}{This is a list of attribute levels, and the type of the variables. D stands for dummy, C stands for continuous variables. Number of attribute levels are introduced through a vector.}

\item{samples}{Number of replications of the design.}

\item{continuous_levels}{Levels of the continuous variable enters separately.}
}
\description{
This function simulates the design of a paper (C Michaud, D Llerena, I Joly, 2013). Six different combinations of the attributes are generated and duplicated to form a questionnaire of 12 possible choices including No-choice alternative. Price is randomly assigned to the alternatives using price_rand function. The design is unlabelled, However; by duplicating the 6 possible pairs of the alternatives, we omitted the biasedness of the data generated using the design.
}
\examples{
input3 = list(attr.lvls = c(2, 2), type = c("D", "D"))
samples = 50
continuous_levels = c(1.5, 2, 2.5, 3, 3.5, 4, 4.5)
paper_design(input3, samples, continuous_levels)
}
