% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{price_rand}
\alias{price_rand}
\title{Randomization of a Continuous Variable}
\usage{
price_rand(N_rows)
}
\arguments{
\item{N_rows}{This parameter should be equal to the number of rows of the design.}
}
\description{
This function samples from levels of a continuous variables such as price. This column will be added to the design. A drawback appears when the number of levels is inadequate.
}
\examples{
N_rows = 28
 price_rand(N_rows)

}
