% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{total_utility}
\alias{total_utility}
\title{Total Utility}
\usage{
total_utility(alt_u_matr, cons_u_indiv, Gumbel = c(0, 1))
}
\arguments{
\item{alt_u_matr}{Alternative specific utility matrix obtained by utility_alt function of the package.}

\item{cons_u_indiv}{Utility matrix obtained given to the socioeconomic characteristics. See utility_characteristics function.}

\item{Gumbel}{A vector to specify the mean and beta parameter of noise generator}
}
\description{
The function sums observed utlity and unobserved utility to compute the total utility for each alternative and individual.
}
\examples{
total_utility(utility_alt(), utility_characteristics(), c(0 ,1))

}
