% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{utility_alt}
\alias{utility_alt}
\title{Alternative Specific Utility}
\usage{
utility_alt(alt_params_matrix, design, interactions = c(2, 3))
}
\arguments{
\item{alt_params_matrix}{Matrix of alternatives' attributes parameters for each individual obtained from alt_params function.}

\item{design}{The design made by one of the design generator functions in this package.}

\item{interactions}{A vector keeping the order of columns of the design for which you wish to have interaction among them.}
}
\description{
This function calculates the alternative's observed utility in which parameters are obtained by alt_params function, and alternatives' attributes are derived from designs made by different functions in SDCM package.
}
\examples{
utility_alt(alt_params(see the function's arguments), FFD(see the function's arguments), interactions = c(2 ,3))

}
