% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{utility_characteristics}
\alias{utility_characteristics}
\title{Utility Obtained from Scoioeconomic Characteristics}
\usage{
utility_characteristics(ind_params_matrix, socioeconomic_data)
}
\arguments{
\item{ind_params_matrix}{It is the matrix of socioeconomic parameters generated by ind_params function in this package.}

\item{socioeconomic_data}{Data synthetically, generated by individual function in this package.}
}
\description{
Socioeconomic characteristics as well as age, gender, habit, and income comprises some part of total utility of a choice made by decision maker. This function, calculates this amount of utility using the characteristics and the assigned parameters.Finally, It will render a 1*N matrix, each column shows the utility gained by each person due to their socioeconomic characteristics.
}
\examples{
utility_characteristics(ind_params(see this function), individual(see this function))

}
