% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wide_to_long}
\alias{wide_to_long}
\title{Converting a Wide Design to a Long Design}
\usage{
wide_to_long(input1, samples)
}
\arguments{
\item{input1}{This is a list containing attribute levels(attr.lvls), continuous varibles' levels(c.lvls) if any exists, and a vector of the type of variables, in which D stands for dummy and C stands for continuous variables.}

\item{samples}{Number of designs to be generated. This option allows the allocation of various designs to individuals. This option will no longer exist once replication option is added.}
}
\description{
This function converts a full factorial design in a wide format to a long format design resulting in a balanced, and orthogonal design. If you want a design which is balanced in choice sets, this function works.
}
\examples{
input1 = list(attr.lvls = c(2, 2, 2, 2), type = c("D", "D", "D", "D"))
 samples = 1000
 wide_to_long(input1, samples)

}
